feat(usage overview): Extract drawer charts into component
